+++ modules/model/src/test/java/net/pantasystem/milktea/model/notes/reaction/ReactionTest.kt
@@ -6,6 +6,60 @@ import org.junit.jupiter.api.Test
 class ReactionTest {
 
     @Test
+    fun isLocal() {
+        val localReaction = Reaction(":kawaii@.:")
+        Assertions.assertTrue(localReaction.isLocal())
+    }
+
+    @Test
+    fun isNotLocal() {
+        val localReaction = Reaction(":kawaii@misskey.io:")
+        Assertions.assertFalse(localReaction.isLocal())
+    }
+
+    @Test
+    fun getName() {
+        val reaction = Reaction(":a:")
+        Assertions.assertEquals("a", reaction.getName())
+    }
+
+    @Test
+    fun getNameThenRemoteCustomEmojiReaction() {
+        val reaction = Reaction(":a@misskey.io:")
+        Assertions.assertEquals("a", reaction.getName())
+    }
+
+    @Test
+    fun getNameThenLocalCustomEmojiReaction() {
+        val reaction = Reaction(":a@.:")
+        Assertions.assertEquals("a", reaction.getName())
+    }
+
+    @Test
+    fun isCustomEmojiFormatThenCustomEmoji() {
+        val reaction = Reaction(":a:")
+        Assertions.assertTrue(reaction.isCustomEmojiFormat())
+    }
+
+    @Test
+    fun isCustomEmojiFormatThenRemoteCustomEmoji() {
+        val reaction = Reaction(":a@misskey.io:")
+        Assertions.assertTrue(reaction.isCustomEmojiFormat())
+    }
+
+    @Test
+    fun isCustomEmojiFormatThenLocalCustomEmoji() {
+        val reaction = Reaction(":a@.:")
+        Assertions.assertTrue(reaction.isCustomEmojiFormat())
+    }
+
+    @Test
+    fun isCustomEmojiFormatThenEmoji() {
+        val reaction = Reaction("😄")
+        Assertions.assertFalse(reaction.isCustomEmojiFormat())
+    }
+
+    @Test
     fun getName_GiveColonNameColon() {
         val reaction = Reaction(":name:")
         Assertions.assertEquals("name", reaction.getName())
@@ -33,5 +87,17 @@ class ReactionTest {
     fun getNameAndHost_GiveColonNameAtMarkHostColon() {
         val reaction = Reaction(":name@host:")
         Assertions.assertEquals("name@host", reaction.getNameAndHost())
+    }
+
+    @Test
+    fun getHost_GiveNameAtMarkDot() {
+        val reaction = Reaction("name@.")
+        Assertions.assertNull(reaction.getHost())
+    }
+
+    @Test
+    fun getNameAndHost_GiveNameAtMarkDot() {
+        val reaction = Reaction("name@.")
+        Assertions.assertEquals("name", reaction.getNameAndHost())
     }
 }
